Rename folder to hyphenated style
