+++ codex-ui/src/vue/components/vertical-menu/VerticalMenu.vue
