+++ codex-ui/src/vue/components/vertical-menu/VerticalMenuItem.vue
